+++ app/src/Functions/removeReference.ts
@@ -2,7 +2,7 @@ import { connectionHandler } from '../Helpers/connection_helpers';
 import { getCurrentItemId, getReferences } from '../Helpers/reference_helpers';
 import _ from 'lodash';
 
-export async function removeReference(collectionId: string, propertyName: string, referringItem: ReferringItem, referencedItem: ReferencedItem, options?: WeivDataOptions): Promise<void> {
+export async function removeReference(collectionId: string, propertyName: string, referringItem: ReferringItem, referencedItem: ReferencedItem, options?: WeivDataOptions): Promise<object> {
     try {
         if (!collectionId || !propertyName || !referringItem || !referencedItem) {
             throw Error(`WeivData - One or more required param is undefined - Required Params: collectionId, propertyName, referringItem, referencedItem`);
@@ -13,16 +13,9 @@ export async function removeReference(collectionId: string, propertyName: string
         const itemId = getCurrentItemId(referringItem);
 
         const { collection, cleanup } = await connectionHandler(collectionId, suppressAuth);
-        const document = await collection.findOne({ _id: itemId }, { readConcern: consistentRead === true ? "majority" : "local" });
-        const isMultiReference = Array.isArray(document?.[propertyName]);
-
-        const updateOperation = isMultiReference
-            ? { $pull: { [propertyName]: { $in: references } }, $set: { _updatedDate: new Date() } }
-            : { $set: { [propertyName]: undefined, _updatedDate: new Date() } };
-
-        const { acknowledged } = await collection.updateOne(
+        const { acknowledged, modifiedCount } = await collection.updateOne(
             { _id: itemId },
-            { ...updateOperation },
+            { $pull: { [propertyName]: { $in: references } }, $set: { _updatedDate: new Date() } },
             { readConcern: consistentRead === true ? "majority" : "local" }
         );
 
@@ -30,8 +23,10 @@ export async function removeReference(collectionId: string, propertyName: string
             await cleanup();
         }
 
-        if (!acknowledged) {
-            throw Error(`WeivData - Error when removing references, acknowledged: ${acknowledged}`)
+        if (!acknowledged || modifiedCount === 0) {
+            throw Error(`WeivData - Error when removing references, acknowledged: ${acknowledged}, modifiedCount: ${modifiedCount}`)
+        } else {
+            return { result: true, updatedCount: modifiedCount };
         }
     } catch (err) {
         throw Error(`WeivData - Error when removing references: ${err}`);