+++ app/src/Functions/replaceReferences.ts
@@ -19,7 +19,7 @@ export async function replaceReferences(collectionId: string, propertyName: stri
         const references = getReferences(referencedItem);
         const itemId = getCurrentItemId(referringItem);
 
-        const updated = await update(collectionId, { _id: itemId, [propertyName]: references.length > 1 ? references : references[0] }, options);
+        const updated = await update(collectionId, { _id: itemId, [propertyName]: references }, options);
 
         if (!updated) {
             throw Error(`WeivData - Error when replacing references, result: ${updated}`)